# models/generic_message.go
// Code generated by go-swagger; DO NOT EDIT.

package models

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	"context"

	"github.com/go-openapi/strfmt"
	"github.com/go-openapi/swag"
)

// GenericMessage GenericMessage generic message
//
// swagger:model GenericMessage
type GenericMessage struct {

	// message
	Message string `json:"message,omitempty"`
}

// Validate validates this generic message
func (m *GenericMessage) Validate(formats strfmt.Registry) error {
	return nil
}

// ContextValidate validates this generic message based on context it is used
func (m *GenericMessage) ContextValidate(ctx context.Context, formats strfmt.Registry) error {
	return nil
}

// MarshalBinary interface implementation
func (m *GenericMessage) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *GenericMessage) UnmarshalBinary(b []byte) error {
	var res GenericMessage
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}